requirements.txt
python-telegram-bot==13.1
googletrans==3.1.0a0